Apply package naming convention: sparkplug -> sparkplug.influxdb
