+++ src/main/java/com/hivemq/extensions/sparkplug/influxdb/metrics/SettableIntGauge.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.extensions.sparkplug.metrics;
+package com.hivemq.extensions.sparkplug.influxdb.metrics;
 
 import com.codahale.metrics.Gauge;
 import com.codahale.metrics.Metric;
@@ -22,32 +22,33 @@ import com.hivemq.extension.sdk.api.annotations.NotNull;
 /**
  * <p>
  * Works like a Gauge, but rather than getting its value from a callback, the value
- * is set when needed.  This can be somewhat convienent, but direct use of a Gauge is likely better
+ * is set when needed.
  * </p>
  * <p>
  * Usage example:
  * <pre>{@code
  *       MetricRegister metricRegistry;
- *       SettableDoubleGauge settable = metricRegistry.register("metric.name", new SettableDoubleGauge());
+ *       SettableLongGauge settable = metricRegistry.register("metric.name", new SettableIntGauge());
  *       // ...
- *       settable.setValue(1.234);
+ *       settable.setValue(100);
  *       // ...
- *       settable.setValue(3.156);
+ *       settable.setValue(200);
  *     }
+ *
  *     </pre>
  */
-public class SettableDoubleGauge implements Metric, Gauge<Double> {
+public class SettableIntGauge implements Metric, Gauge<Integer> {
     /**
      * Current value.  Assignment will be atomic.  <a href="http://docs.oracle.com/javase/specs/jls/se7/html/jls-17.html#jls-17.7">See 17.7</a>
      */
-    private volatile double value;
+    private volatile int value = 0;
 
     /**
-     * The last value set by {@link #setValue(double)}}
+     * The last value set by {@link #setValue(int)}}
      *
      * @return Last set value, or zero.
      */
-    public @NotNull Double getValue() {
+    public @NotNull Integer getValue() {
         return value;
     }
 
@@ -57,7 +58,7 @@ public class SettableDoubleGauge implements Metric, Gauge<Double> {
      * @param value last set value
      * @return itself
      */
-    public @NotNull SettableDoubleGauge setValue(final double value) {
+    public @NotNull SettableIntGauge setValue(final int value) {
         this.value = value;
         return this;
     }
