+++ src/test/java/com/hivemq/extensions/sparkplug/influxdb/configuration/PropertiesReaderTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.extensions.sparkplug.configuration;
+package com.hivemq.extensions.sparkplug.influxdb.configuration;
 
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import org.junit.jupiter.api.Test;
