+++ src/test/java/com/hivemq/extensions/sparkplug/influxdb/topics/TopicStructureTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.extensions.sparkplug.topics;
+package com.hivemq.extensions.sparkplug.influxdb.topics;
 
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import org.junit.jupiter.api.Test;
